Version 2026.1.25
Patch général et résolutions des problèmes connus. Les données sont maintenant sauvegardées correctement.
La commande "set_minimum_age" a été changée en "set-minimum-age"
La commande "welcome-set" devrait à présent fonctionner
⚠️ Aucun test n'a été fait avec cette nouvelle version, il se pourrait que certaines fonctionnalitées ne soient pas encore réparées

+++ Files/Modules/Main/help_command.py
@@ -115,7 +115,10 @@ class HelpCommand(commands.Cog):
         # Supprimer le message après 30 secondes d'inactivité
         await asyncio.sleep(30)
         try:
-            await message.delete()
+            if 'message' in locals():
+                await message.delete()
+            else:
+                await interaction.response.send_message("Aucun message à supprimer.", ephemeral=True)
         except discord.NotFound:
             pass
